requirements.txt
flask
opencv-python
psycopg2-binary
python-dotenv
